Moved commandDetails
