+++ src/commandDetails/coin/adjust.ts
@@ -67,19 +67,19 @@ export const coinAdjustCommandDetails: CodeyCommandDetails = {
   options: [
     {
       name: 'user',
-      description: 'The user to adjust the balance of,',
+      description: 'The user to adjust the balance of.',
       type: CodeyCommandOptionType.USER,
       required: true,
     },
     {
       name: 'amount',
-      description: 'The amount to adjust the balance of the specified user to,',
+      description: 'The amount to adjust the balance of the specified user by.',
       type: CodeyCommandOptionType.NUMBER,
       required: true,
     },
     {
       name: 'reason',
-      description: 'The reason why we are adjusting the balance,',
+      description: 'The reason why we are adjusting the balance.',
       type: CodeyCommandOptionType.STRING,
       required: false,
     },
