+++ src/commands/coin/coin.ts
@@ -3,13 +3,13 @@ import { CodeyCommand, CodeyCommandDetails } from '../../codeyCommand';
 import { coinAdjustCommandDetails } from '../../commandDetails/coin/adjust';
 import { coinCheckCommandDetails } from '../../commandDetails/coin/check';
 import { coinInfoCommandDetails } from '../../commandDetails/coin/info';
-import { coinUpdateCommandDetails } from '../../commandDetails/coin/update';
 import { coinCurrentLeaderboardCommandDetails } from '../../commandDetails/coin/leaderboard';
+import { coinUpdateCommandDetails } from '../../commandDetails/coin/update';
 
 const coinCommandDetails: CodeyCommandDetails = {
   name: 'coin',
   aliases: [],
-  description: 'Handles coin functions',
+  description: 'Handle coin functions.',
   detailedDescription: `**Examples:**
 \`${container.botPrefix}coin adjust @Codey 100\`
 \`${container.botPrefix}coin adjust @Codey -100 Codey broke.\`
