+++ src/utils/num.ts
@@ -1,0 +1,3 @@
+export const getRandomInt = (max: number): number => {
+  return Math.floor(Math.random() * max) + 1;
+};
